moved functions and interface from EditComputeDataset component
