+++ src/@types/ComputeDataset.ts
@@ -1,0 +1,4 @@
+export interface AlgorithmOption {
+  did: string
+  name: string
+}
